refactor: extract program tab to views/ProgramView.jsx
Strangler-fig extraction #5 of the sprint — the single largest seam
(~3,071-line program tab, #77). PR 1 of the program decomposition: moves
the whole tab out first; PR6/PR7 will split its week/year sub-tabs into
views/program/*. Behaviour unchanged.

- new views/ProgramView.jsx owns local progTab state (phases | week |
  year) + the sub-nav; takes expanded/setExpanded as props
- program-structure reference data (ERGZONE_FORMAT, ANNUAL_ARC,
  EVENT_PATHWAY, BUILD1_SESSIONS, technique/mobility drills, conversions
  — 13 consts) moves to the new constants/program.js (pure data, excluded
  from coverage)
- monolith renders {view === 'program' && <ProgramView .../>} and sheds
  ~15 now-orphaned constant imports (exercises + nutrition fully)
- render test covers all three sub-tabs (measured 89.8/92.5 lines/funcs,
  69 branches — large presentational glue, gated via the global floor
  like RecoveryView, not a per-file 80/80/70)

Monolith: 4,455 -> 963 lines.
